+++ R/parse-preref.R
@@ -3,6 +3,9 @@ parse_preref <- function(x, file, offset = x[[1]]) {
   if (length(tags) == 0)
     return()
 
+  ## Switch markdown on/off if md absent/present
+  markdown_on("md" %in% vapply(tags, "[[", "", "tag"))
+
   tags <- parse_description(tags)
   tags <- compact(lapply(tags, parse_tag))
 
@@ -73,6 +76,16 @@ parse_description <- function(tags) {
 }
 
 # Individual tag parsers --------------------------------------------------
+
+parse.value.markdown <- function(x) {
+  x$val <- full_markdown(x$val)
+  parse.value(x)
+}
+
+parse.value.restricted.markdown <- function(x) {
+  x$val <- restricted_markdown(x$val)
+  parse.value(x)
+}
 
 # TODO: move into own file tag-parsers.R
 # TODO: consistent naming scheme `parse.value` -> `tag_value`
@@ -147,7 +160,13 @@ parse.words.line <- function(x) {
   }
 }
 
-parse.name.description <- function(x) {
+parse.name.description.nomarkdown <- function(x) {
+  parse.name.description(x, markdown = FALSE)
+}
+
+parse.name.description <- function(x, markdown = TRUE) {
+  if (markdown) x$val <- full_markdown(x$val)
+  
   if (x$val == "") {
     tag_warning(x, "requires a value")
   } else if (!str_detect(x$val, "[[:space:]]+")) {