+++ tests/testthat/test-examples.R
@@ -3,9 +3,9 @@ context("Examples")
 test_that("@example loads from specified files", {
   out <- roc_proc_text(rd_roclet(), "
     #' @name a
+    #' @title a
     #'
     #' @example Rd-example-1.R
-    #'
     #' @example Rd-example-2.R
     NULL")[[1]]
 
@@ -17,6 +17,7 @@ test_that("@example loads from specified files", {
 test_that("@examples captures examples", {
   out <- roc_proc_text(rd_roclet(), "
     #' @name a
+    #' @title a
     #'
     #' @examples a <- 2
     #'
@@ -29,6 +30,7 @@ test_that("@examples captures examples", {
 test_that("@examples and @example combine", {
   out <- roc_proc_text(rd_roclet(), "
     #' @name a
+    #' @title a
     #' @example Rd-example-1.R
     #' @examples a <- 2
     NULL")[[1]]
@@ -41,6 +43,7 @@ test_that("@examples and @example combine", {
 test_that("@example does not introduce extra empty lines", {
   out <- roc_proc_text(rd_roclet(), "
     #' @name a
+    #' @title a
     #' @example Rd-example-3.R
     NULL")[[1]]
 
@@ -52,6 +55,7 @@ test_that("@example gives warning if used instead of @examples", {
   expect_warning(
     out <- roc_proc_text(rd_roclet(), "
       #' @name a
+      #' @title a
       #' @example
       #' a <- 1
       #' a + b
@@ -66,6 +70,7 @@ test_that("@example gives warning if used instead of @examples", {
 test_that("indentation in examples preserved", {
   out <- roc_proc_text(rd_roclet(), "
     #' @name a
+    #' @title a
     #' @examples a <-
     #'     2
     NULL")[[1]]
@@ -77,6 +82,7 @@ test_that("indentation in examples preserved", {
 test_that("% and \\ in @example escaped", {
   out <- roc_proc_text(rd_roclet(), "
     #' @name a
+    #' @title a
     #' @example Rd-example-4.R
     NULL")[[1]]
 
@@ -87,6 +93,7 @@ test_that("% and \\ in @example escaped", {
 test_that("\\dontrun in @example unescaped", {
   out <- roc_proc_text(rd_roclet(), "
     #' @name a
+    #' @title a
     #' @example Rd-example-5.R
     NULL")[[1]]
 
@@ -97,6 +104,7 @@ test_that("\\dontrun in @example unescaped", {
 test_that("% in @examples escaped before matching braces test (#213)", {
   out <- roc_proc_text(rd_roclet(), "
     #' @name a
+    #' @title a
     #' @examples
     #' {a %% b}
     NULL")[[1]]
@@ -108,6 +116,7 @@ test_that("% in @examples escaped before matching braces test (#213)", {
 test_that("multiple examples (#470)", {
   out <- roc_proc_text(rd_roclet(), "
     #' @name a
+    #' @title a
     #' @examples
     #' TRUE
     #' @examples