+++ tests/testthat/test-keyword.R
@@ -3,6 +3,7 @@ context("Keyword")
 test_that("keywords split into pieces", {
   out <- roc_proc_text(rd_roclet(), "
     #' @keywords a b
+    #' @title a
     #' @name a
     NULL")[[1]]