+++ tests/testthat/test-s4.R
@@ -53,3 +53,14 @@ test_that("setMethod equivalent to setReplaceMethod", {
 
   expect_equal(out[[2]]$object, out[[3]]$object)
 })
+
+test_that("trailing arguments get signature ANY", {
+  out <- roc_proc_text(rd_roclet(), "
+    setGeneric('foo', function(x, y) standardGeneric('foo'))
+
+    #' foo
+    setMethod('foo', c('numeric', 'ANY'), function(x, y) x)
+  ")[[1]]
+
+  expect_equal(out$get_tag("alias")$values, "foo,numeric,ANY-method")
+})