+++ NEWS.md
@@ -1,5 +1,9 @@
 # roxygen2 5.0.1.9000
 
+* Most fields can now be written using Markdown markup instead of the
+  traditional Rd language. See the 'markdown' vignette for details
+  (#364, #431), by @gaborcsardi
+
 # roxygen2 5.0.1
 
 * Use `ls()`, not `names()` to list elements of environment: fixes R 3.1.0