+++ NEWS.md
@@ -1,5 +1,8 @@
 # roxygen2 5.0.1.9000
 
+* Fixed an issue where `.`s were sometimes added between words within
+  a `@family` tag (#477, @kevinushey).
+
 # roxygen2 5.0.1
 
 * Use `ls()`, not `names()` to list elements of environment: fixes R 3.1.0