+++ NEWS.md
@@ -3,6 +3,8 @@
 * Fixed an issue where `.`s were sometimes added between words within
   a `@family` tag (#477, @kevinushey).
 
+* Multiple `@examples` sections are merged (#472, @krlmlr).
+
 # roxygen2 5.0.1
 
 * Use `ls()`, not `names()` to list elements of environment: fixes R 3.1.0