+++ NEWS.md
@@ -1,6 +1,6 @@
 # roxygen2 5.0.1.9000
 
-* Specail characters in `@describeIn` function names are escaped (#450).
+* Special characters in `@describeIn` function names are escaped (#450).
 
 * Ensure that `functions` with S3 class are still treated as functions (#455).