+++ NEWS.md
@@ -1,5 +1,8 @@
 # roxygen2 5.0.1.9000
 
+* Non-primitive, internal S3 generics (e.g. 'rbind', 'cbind') are now properly
+  detected as S3 generics. (#488, @kevinushey)
+
 * Changes to DESCRIPTION (i.e. `Collate:` and `RoxygenNote`) now use
   the desc package. This will minimise spurious changes (#430).